requirements.txt
kafka-python
requests
python-dotenv
faker
snowflake-connector-python